Addressed code-related changes from PR feedback
