+++ .prettierignore
@@ -2,5 +2,3 @@
 
 /dist
 /coverage
-
-apps/ownyourday-client/vite.config.ts
