+++ apps/ownyourday-client/src/components/app-routes/index.ts
@@ -1,0 +1,1 @@
+export * from './AppRoutes';
