+++ apps/ownyourday-client/src/routes.ts
@@ -1,0 +1,2 @@
+export const SIGN_IN = '/sign-in';
+export const TODAY = '/today';
