+++ package.json
@@ -12,7 +12,6 @@
         "@emotion/react": "^11.10.6",
         "@emotion/styled": "^11.10.6",
         "aws-amplify": "^5.0.19",
-        "axios": "^1.3.4",
         "express": "^4.18.1",
         "framer-motion": "^10.6.0",
         "react": "18.2.0",
